installer: Add support for insecure_installation
This patch adds kernel command line argument: 'insecure_installation'
as a flag that user can set to 1 to allow some operations that aren't
normally allowed due to security concerns. This is disabled by default
and it is up to the user to the ensure security when this options is
enabled.

This patch also disables by default download of ks_config file from non-HTTPS
sources.

Following cases can arise:
1) insecure_installation=1 passed as kernel param in cmdline
   insecure_installation key if present in ks_config ignored
   - ks_config and rpms from untrusted sources -> allowed
2) insecure_installation=0 passed as kernel param in cmdline
   insecure_installation key if present in ks_config ignored
   - ks_config and rpms from untrusted sources -> not allowed
3) insecure_installation not passed as kernel param in cmdline
     insecure_installation: true passed in ks_config
   - ks_config from http -> not allowed
   - rpms from untrusted sources -> allowed
4) insecure_installation not passed as kernel param in cmdline
     insecure_installation: false passed in ks_config
   - ks_config from http -> not allowed
   - rpms from untrusted sources -> not allowed

We do sslverify=0 to download rpms from untrusted sources too.
We pass insecure_installation key from isoInstaller.py to installer.py in
install_config.
In the case when insecure_installation is passed in command line but no
ks_config is provided, meaning install_config passed to installer.py should
be empty. Hence to accomodate this we always pop the insecure_installaion key
first if present and add it at later point.