m_{postinstall,preinstall}.py: Fix post install arbitrary script execution
Post and pre install scripts are listed using os.listdir() after copying
to temp dir which doesn't confirm the order of execution of scripts
as provided in kickstart.

Current change maintains a seperate list of scripts with their order
of execution and checks if the script is executable before running it.

If not executable script execution is skipped.

Change-Id: Ia6312330ce919f389994e6ffc06bd87de8a9ffaa