Add yaml parser for ks config
This add a yaml parser, so the kickstart file can be in yaml format. Additionally,
this add the !param tag to the yaml config. This allows paramters to be set on the
commandd line. For example, to set the "disk" setting:

disk: !param device

This way the "disk" needs to be set on the command line, with:

photon-installer ... args ... --param device=/dev/loop7

Defaults can also be set, which makes setting the parameters optional.
Example:

eject_cdrom: !param eject=false

Change-Id: I0beeaab68dcd448a7c0e64844bebb8907412245c