handle LVM on the host
The installer threw errors when LVM logival volume groups
are in use on the host, becausev p-o-i tries to clear all
of them. This change clears LGs only for those we need.

Also check if LG names in config are the same as on the host
and bail in that case . These *may* be left over from a previous
failed install, in that case suggest to remove them manually.

Change-Id: I7f1fe89991208234fcdf904d6901133f9e92dca7