sample_ks: Fix the field length to get the disk name
The Issue was fetching a specifc field to get the disk name but the space is vary based on size of file.
Like sometime the size can be in a single digit in that case the same field will give different parameter and test got failed.
Hence replacing cut with awk.

Change-Id: Ic24747efffc63cedea9c45acc00d4250096553a2