sample_ks: Fix the field length to get the disk name
The Issue was fetching a specifc field to get the disk name but the space is vary based on size of file.
Like sometime the size can be in a single digit in that case the same field will give different parameter and test got failed.
Hence modifying the on the disk command.

Change-Id: Ide7f76ac35e6985407ed9251522fd4e93cde0f58