installer:modules: Fix module execute
- In preinstall phase, self.progress_bar object
is unavailable. so the progress_bar log causes failure
when 'ui' is set to true and `preinstall` is defined.
Issue can be reproduced using below kickstart file:
https://github.com/vmware/photon-os-installer/blob/master/sample_ks/sample_ks_preinstall_postinstall.cfg
Additionally, i observe that update of "Setting up GRUB"
was redundant and not required during module execution.

Same is already present in _setup_grub() call.
https://github.com/vmware/photon-os-installer/blob/master/photon_installer/installer.py#L1333

To fix the crash, removed redundant update message.

Signed-off-by: Ankit Jain <ankit-aj.jain@broadcom.com>